Add necessary model and service

+++ src/main/java/edu/neu/cs6510/pnnl/hunting/model/Ahu1Thresholds.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class Ahu1Thresholds {
     private Integer id;
 
-    private Integer minimumOutdoorAirFlowSetPoint;
+    private Double minimumOutdoorAirFlowSetPoint;
 }
