+++ src/main/java/edu/neu/cs6510/pnnl/hunting/model/Ahu3.java
@@ -10,7 +10,7 @@ public class Ahu3 {
 
     private Double exhaustAirFlow;
 
-    private Integer minimumOutdoorAirFlow;
+    private Double minimumOutdoorAirFlow;
 
     private Double outdoorAirFlow;
 }
